requirements.txt
dnspython==2.3.0
requests==2.27.1
python-dotenv==0.19.2